+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 ## Changelog
 - Update 1.21.8
+- Fixed compass widget spawning off screen
+- Fixed coordinates display integration applying incorrect position modifications
+- Fixed arrow widget not rendering count
+- Uses custom color picker screen